[#5] refactor: repository package 생성
- RestaurantRepository 패키지 이동
