+++ src/main/java/dev/grabros/osikdang/repository/RestaurantRepository.java
@@ -1,5 +1,6 @@
-package dev.grabros.osikdang.domain.restaurant;
+package dev.grabros.osikdang.repository;
 
+import dev.grabros.osikdang.domain.restaurant.Restaurant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RestaurantRepository extends JpaRepository<Restaurant, Long> {
